+++ packages/create-mastra/CHANGELOG.md
@@ -1,5 +1,9 @@
 # create-mastra
 
+## 1.1.0-alpha.2
+
+## 1.1.0-alpha.1
+
 ## 1.1.0-alpha.0
 
 ### Patch Changes